requirements.txt
streamlit
fitz
python-docx
requests
beautifulsoup4
sentence-transformers
scikit-learn
pandas
tqdm
